build: support (and require) Automake-NG
* configure.ac (AC_PREREQ): Require Autoconf version >= 2.65,
since that is the minimal version supported by Automake-NG.
(AM_INIT_AUTOMAKE): Add the 'ng' option, to ensure that mainstream
Automake is not used by mistake when bootstrapping.  Also, bump the
required Automake version from '1.11.1' to '1.12a', which is the
latest (and still development-only) version of Automake-NG at the
moment of writing.
* bootstrap.conf ($buildreq): Require 'automake-ng' and
'aclocal-ng'; don't require mainstream 'automake' anymore.
Bump required 'autoconf' version to 2.65.
* m4/mkdirp-compat.m4: New file, contain a definition of the macro
'AM_PROG_MKDIR_P' (simply as an alias to the Autoconf-provided
macro 'AC_PROG_MKDIR_P'), that is used by the Gettext-provided
macro 'AM_PO_SUBDIRS', but which has been removed in Automake-NG
(as well as in the master branch of mainline Automake).

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>